Combine client files into one single file
